🔒token + auth

+++ src/controllers/ArticleController.js
@@ -19,52 +19,67 @@ const getArticle = async (req, res) => {
 };
 
 const creatArticle = async (req, res) => {
+  const isAdmin = req.user.isAdmin;
   const { title_Produit, imageURL, price, description } = req.body;
   try {
+    if (isAdmin) {
+      const article = new Article({
+        title_Produit,
+        imageURL,
+        price,
+        description,
+      });
+      console.log(article);
+      await article.save();
+      res.json({ message: "Article created", article });
+    } else {
+      res.status(401).send("non-authorized");
+    }
     // const imageURL = req.file.filename;
-    const article = new Article({
-      title_Produit,
-      imageURL,
-      price,
-      description,
-    });
-    console.log(article);
-    await article.save();
-    res.json({ message: "Article created", article });
   } catch (error) {
     res.status(400).json({ message: "error create article" });
   }
 };
 
 const updateArticle = async (req, res) => {
+  const isAdmin = req.user.isAdmin;
   const { id } = req.params;
   const { title_Produit, imageURL, price, description } = req.body;
   try {
-    const updatedArticle = await Article.findByIdAndUpdate(
-      id,
-      {
-        title_Produit,
-        imageURL,
-        price,
-        description,
-      },
-      { new: true }
-    );
-    res.json({ message: "Article updated", updatedArticle });
+    if (isAdmin) {
+      const updatedArticle = await Article.findByIdAndUpdate(
+        id,
+        {
+          title_Produit,
+          imageURL,
+          price,
+          description,
+        },
+        { new: true }
+      );
+      res.json({ message: "Article updated", updatedArticle });
+    } else {
+      res.status(401).send("non-authorized");
+    }
   } catch (error) {
     res.status(400).json({ message: "error update article" });
   }
 };
 
 const deleteArticle = async (req, res) => {
+  const isAdmin = req.user.isAdmin;
   const articleId = req.params.id;
   try {
-    const article = await Article.findByIdAndDelete(articleId);
-    if (!article) {
-      // Vérifiez si aucun article n'a été trouvé avec cet ID
-      return res.status(404).json({ message: "Article not found" });
+    if (isAdmin) {
+      const article = await Article.findByIdAndDelete(articleId);
+      if (!article) {
+        // Vérifiez si aucun article n'a été trouvé avec cet ID
+        return res.status(404).json({ message: "Article not found" });
+      }
+      res.json(article);
+    } else {
+      es.status(401).send("non-authorized");
     }
-    res.json(article);
   } catch (err) {
     console.error(err);
     res.status(400).json({ message: " error delete" });
